backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
transformers==4.39.3
tokenizers==0.15.2
sentence-transformers==2.2.2
faiss-cpu==1.9.0
numpy==1.26.4
python-multipart==0.0.6
torch>=2.0.0
streamlit
requests
pandas
plotly